Clean up unused code, rename stuff to be in line with repo practices
