+++ norbit-webapp/components/SensorConfigurationPanel/SensorConfigurationPanel.tsx
