+++ norbit-webapp/components/SensorConfigurationPanel/types.ts
@@ -1,4 +1,4 @@
-import { TamplifyInstance } from "@/dashboard/Dashboard";
+import { TamplifyInstance } from "@/Dashboard/Dashboard";
 
 export interface ConfigProps {
   amplifyInstance?: TamplifyInstance | null;
