+++ norbit-webapp/components/SidebarComponent/types.tsx
@@ -1,4 +1,4 @@
-import { TamplifyInstance } from "@/dashboard/Dashboard";
+import { TamplifyInstance } from "@/Dashboard/Dashboard";
 
 export interface SidebarProps {
   amplifyInstance?: TamplifyInstance | null;
