+++ norbit-webapp/pages/index.tsx
@@ -1,6 +1,6 @@
 import Head from "next/head";
 import styles from "../styles/Home.module.css";
-import Dashboard from "components/dashboard/Dashboard";
+import Dashboard from "@/Dashboard/Dashboard";
 
 export default function Home() {
   return (
